requirements.txt
streamlit>=1.35.0
chromadb>=0.5.0
huggingface-hub>=0.30.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
pypdf>=4.0.0
pandas>=2.0.0
